Initial modular structure and compiler
