+++ README.md
@@ -1,5 +1,4 @@
 # LWO Loader Tests
-
 Test models and automated testing environment for the [three.js LWOLoader](https://threejs.org/examples/#webgl_loader_lwo).
 
 If you want to use the LWOLoader in your projects, get the [latest version from the main three.js repo](https://github.com/mrdoob/three.js/blob/dev/examples/js/loaders/LWOLoader.js).
@@ -10,3 +9,10 @@ Licensed under Creative Commons Attribution 4.0 International License
 
 Environment images by [HDRI Haven](https://hdrihaven.com).
 Licensed under Creative Commons CC0 Public Domain
+
+# Development & Testing
+`npm install` to install development dependencies.
+
+
+`npm run build` to build LWOLoader bundled file.
+`npm run test` to build and initialize http-server.
