+++ index.html
@@ -17,7 +17,7 @@
 		<script src="https://cdnjs.cloudflare.com/ajax/libs/stats.js/r16/Stats.min.js" integrity="sha256-eQxIuR9Z5Afu5SFVtoND4dHdHygFRhueTk7mwkINjro=" crossorigin="anonymous"></script>
 		<script src="https://unpkg.com/three@0.103.0/examples/js/controls/OrbitControls.js" integrity="sha384-A8KifjKpxN1LUAWjByefDXyOrSL2gDQXe3TgnEL51gOuikIFuchQzDYWUNKdfE6A" crossorigin="anonymous"></script>
 		<script src="https://cdnjs.cloudflare.com/ajax/libs/dat-gui/0.7.6/dat.gui.min.js" integrity="sha256-HJ7j+71YYw6Kcs8THwQV9lXmPOcR0eXlg7n8KRTZsyA=" crossorigin="anonymous"></script>
-		<script src="src/LWOLoader.js" crossorigin="anonymous"></script>
+		<script src="build/LWOLoader.js" crossorigin="anonymous"></script>
 
 	</head>
